Read file content from S3 bucket with boto3
In this function, bucket_name is the name of your S3 bucket, and file_name is the name of the file you want to read. The get_object method returns a dictionary that contains the file data in the 'Body' key. We then read this data using the read method.